% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_factor.R
\name{discrim_factors}
\alias{discrim_factors}
\title{Identify factors discriminating clusters}
\usage{
discrim_factors(data, groups, p.adj = "none")
}
\arguments{
\item{data}{: a matrix of features (with features as column)}

\item{groups}{: a vector of groups binning the matrix of features}

\item{p.adj}{: The method to correct the `p-value`. See \link[stats]{p.adjust} for available methods.}
}
\value{
a dataframe containing the cumulated LDA scaling scores of each feature, the mean and SD of each feature for each group, and the p-value of the ANOVA computed with each feature as response variable, and the grouping vector as explanatory variable.
}
\description{
Identify factors discriminating clusters
}
\details{
This function seeks to measure the importance of each variable on the composition of clusters. It computes a \link[MASS]{lda} and extracts the scaling as proxy of the importance of the variable on cluster separation.
}
\examples{
data(iris)
my_features <- iris[1:4]
my_groups <- iris$Species

discrim_factors(my_features, my_groups, "bon")

}
