% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_factors_pairwise.R
\name{discrim_factors_pairwise}
\alias{discrim_factors_pairwise}
\title{Compute discriminating features betyeen cluster}
\usage{
discrim_factors_pairwise(data, groups, p.adj = "bon")
}
\arguments{
\item{data}{a matrix or dataframe of features}

\item{groups}{a vector of groups}

\item{p.adj}{The method to correct the `p-value`. See \link[stats]{p.adjust} for available methods.}
}
\value{
a named list of pairwise comparisond
}
\description{
This function seeks to measure the importance of each variable on the composition of clusters. It computes a \link[MASS]{lda} and extracts the scaling as proxy of the importance of the variable on cluster separation.
}
\examples{
#' data(iris)
my_features <- iris[1:4]
my_groups <- iris$Species

discrim_factors_pairwise(my_features, my_groups, "bon")

}
