% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_pca_plot.r
\name{pretty_pca_plot}
\alias{pretty_pca_plot}
\title{Pretty PCA Plot}
\usage{
pretty_pca_plot(obj, month, year, type = "default", ...)
}
\arguments{
\item{obj}{A PCA object computed with prcomp or princomp}

\item{month}{an ordered vector with the months}

\item{year}{an ordered vector with the year}

\item{type}{`default` or `ggplot` to decide which kind of graph you want to do}

\item{...}{optional parameters to be passed to the plot function}
}
\value{
a beautiful plot
}
\description{
Pretty PCA Plot
}
\examples{

month <- rep(month.name, 3)
year <- c(rep(2014, 12), rep(2015, 12), rep(2016, 12))

data <- matrix(ncol = 10,
               nrow = 36,
               data = rnorm(360))

data.pca <- prcomp(data)
pretty_pca_plot(data.pca, month, year, main = "test plot")

}
