% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_dates.R
\name{rearrange_dates}
\alias{rearrange_dates}
\title{Rearrange dates}
\usage{
rearrange_dates(dates)
}
\arguments{
\item{dates}{a character vector of dates}
}
\value{
a data.frame containing the dates in ISO format, years, months, and days
}
\description{
This function takes as input a vector of characters representing dates, (i.e. "2015-9-28"), in any format, converts
it into iso time and split years, months, and day
}
\examples{
year <- c(rep(2014,12), rep(2015, 12), rep(2016, 12))
month <- rep(1:12, 3)
day <- sample(1:28, 36, replace = TRUE)
my_date <- paste(year, month, day, sep="-")

rearrange_dates(my_date)
}
